fix: floor model file renamed, fix grid floor store reference
