+++ src/components/Settings/FloorSettings.vue
@@ -142,7 +142,7 @@
 
 <script lang="ts">
 import { defineComponent } from "vue";
-import { Floor } from "@/models/printer-floor/printer-floor.model";
+import { Floor } from "../../models/floors/floor.model";
 import { usePrinterStore } from "../../store/printer.store";
 import { useDialogsStore } from "@/store/dialog.store";
 import { DialogName } from "@/components/Generic/Dialogs/dialog.constants";
