+++ src/event-bus/socketio.events.ts
@@ -1,4 +1,4 @@
-export const sseMessageGlobal = "sse-message-global";
-export const socketIoFloors = "sse-floors";
+export const sseMessageGlobal = "socketio-message-global";
+export const socketIoFloors = "socketio-floors";
 export const sseTestPrinterUpdate = (correlationToken: string) =>
   `sse-message-test-printer-${correlationToken}`;
